requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   openai
    #   starlette
asgi-lifespan==2.1.0
    # via MHChatbot (pyproject.toml)
async-timeout==4.0.3
    # via asyncpg
asyncpg==0.30.0
    # via MHChatbot (pyproject.toml)
bcrypt==4.2.1
    # via MHChatbot (pyproject.toml)
beanie==1.27.0
    # via MHChatbot (pyproject.toml)
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   beanie
    #   uvicorn
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via
    #   email-validator
    #   pymongo
ecdsa==0.19.0
    # via python-jose
email-validator==2.2.0
    # via pydantic
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.6
    # via MHChatbot (pyproject.toml)
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jiter==0.8.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==0.3.21
    # via
    #   MHChatbot (pyproject.toml)
    #   langchain-openai
    #   langgraph
    #   langgraph-checkpoint
langchain-openai==0.2.11
    # via MHChatbot (pyproject.toml)
langgraph==0.2.54
    # via MHChatbot (pyproject.toml)
langgraph-checkpoint==2.0.8
    # via langgraph
langgraph-sdk==0.1.43
    # via langgraph
langsmith==0.1.147
    # via langchain-core
lazy-model==0.2.0
    # via beanie
motor==3.6.0
    # via beanie
msgpack==1.1.0
    # via langgraph-checkpoint
numpy==2.1.3
    # via pandas
openai==1.56.1
    # via
    #   MHChatbot (pyproject.toml)
    #   langchain-openai
orjson==3.10.12
    # via
    #   langgraph-sdk
    #   langsmith
packaging==24.2
    # via langchain-core
pandas==2.2.3
    # via MHChatbot (pyproject.toml)
passlib==1.7.4
    # via MHChatbot (pyproject.toml)
psycopg2-binary==2.9.9
    # via MHChatbot (pyproject.toml)
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pydantic[email]==2.10.3
    # via
    #   MHChatbot (pyproject.toml)
    #   beanie
    #   fastapi
    #   langchain-core
    #   langsmith
    #   lazy-model
    #   openai
    #   pydantic-settings
pydantic-core==2.27.1
    # via pydantic
pydantic-settings==2.6.1
    # via MHChatbot (pyproject.toml)
pymongo==4.9.2
    # via motor
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   MHChatbot (pyproject.toml)
    #   pydantic-settings
python-jose==3.3.0
    # via MHChatbot (pyproject.toml)
python-multipart==0.0.6
    # via MHChatbot (pyproject.toml)
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via langchain-core
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9
    # via python-jose
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   asgi-lifespan
    #   openai
sqlalchemy==2.0.36
    # via MHChatbot (pyproject.toml)
starlette==0.41.3
    # via
    #   MHChatbot (pyproject.toml)
    #   fastapi
tenacity==9.0.0
    # via langchain-core
tiktoken==0.8.0
    # via langchain-openai
toml==0.10.2
    # via beanie
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   beanie
    #   fastapi
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   uvicorn
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
uvicorn==0.23.2
    # via MHChatbot (pyproject.toml)
